+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 ### What I want to add
 
 - Fix text added to places ONLY when places are rear the user position (i.e. for places in front of user initial position, places have only icon, otherwise text+icon)
-- Remove every setInterval interaction between aframe components
 - Add an alert also when user has not granted permissions for geolocation (for the current webapp)
 - Do not update entity (place) position if there is too much difference between consequential user GPS positions (i.e. gps signal is not so stable, so if position is changing a lot in 1-2 seconds, do not consider that --> this may cause problems if user is in a car or some kind of fast moving transport?)
 - Add UX/UI and messages when gps data is not available or there's very poor signal (is there any events for that?)
@@ -20,6 +19,7 @@
 - Added a better demo
 - Enhanced Documentation
 - Added an alert message when geolocation is switched off (from Phone Settings)
+- Remove every setInterval interaction between aframe components
 
 
 ### About demos..